move createTransferOwnershipTx as it is EVM specific
